+++ app/controllers/all/opening/projects_controller.rb
@@ -13,6 +13,19 @@ class All::Opening::ProjectsController < ApplicationController
     @date = "#{Date::MONTHNAMES[month.to_i]} #{year}"
   end
 
+  def download_csv
+    authorize Project, :index?
+
+    @projects = ProjectsFetcher.new.sorted_openers(month, year)
+    @date = "#{Date::MONTHNAMES[month.to_i]} #{year}"
+    @month = month
+    @year = year
+
+    @csv = OpeningProjectsCsvExporter.new(@projects).call
+
+    send_data @csv, filename: "opening_#{month}_#{year}.csv", type: :csv, disposition: "attachment"
+  end
+
   private def month
     params[:month]
   end